+++ src/main/kotlin/com/estapar/domain/car/logging/entry/CarEntryService.kt
@@ -1,4 +1,4 @@
-package com.estapar.domain.car.entry
+package com.estapar.domain.car.logging.entry
 
 import com.estapar.domain.car.logging.AlreadyHasActiveGarageLoggingException
 import com.estapar.domain.car.logging.GarageLoggingService
