+++ src/main/kotlin/com/estapar/infrastructure/api/dto/GarageEntryEvent.kt
@@ -1,18 +1,18 @@
-package com.estapar.infrastructure.api
+package com.estapar.infrastructure.api.dto
 
-import com.estapar.domain.car.exit.CarExit
+import com.estapar.domain.car.logging.entry.CarEntry
 import com.fasterxml.jackson.annotation.JsonProperty
 import java.time.LocalDateTime
 
-open class GarageExitEvent(
-    @JsonProperty("exit_time") val exitTime: LocalDateTime,
+open class GarageEntryEvent(
+    @JsonProperty("entry_time") val entryTime: LocalDateTime,
     @JsonProperty("license_plate") licensePlate: String,
     @JsonProperty("event_type") eventType: GarageEventType
 ) : GarageWebhookEvent(licensePlate, eventType) {
 
-    fun toDomain(): CarExit =
-        CarExit(
-            exitTime = exitTime,
+    fun toDomain(): CarEntry =
+        CarEntry(
+            entryTime = entryTime,
             licensePlate = licensePlate)
 
 }
