+++ src/main/kotlin/com/estapar/infrastructure/api/dto/GarageEventType.kt
@@ -1,4 +1,4 @@
-package com.estapar.infrastructure.api
+package com.estapar.infrastructure.api.dto
 
 enum class GarageEventType {
     ENTRY,
