+++ src/main/kotlin/com/estapar/infrastructure/component/CarEntryServiceImpl.kt
@@ -1,6 +1,6 @@
 package com.estapar.infrastructure.component
 
-import com.estapar.domain.car.entry.CarEntryService
+import com.estapar.domain.car.logging.entry.CarEntryService
 import com.estapar.domain.car.logging.GarageLoggingService
 import org.springframework.stereotype.Service
