+++ src/main/kotlin/com/estapar/infrastructure/component/CarExitServiceImpl.kt
@@ -1,7 +1,7 @@
 package com.estapar.infrastructure.component
 
 import com.estapar.domain.car.billing.BillingService
-import com.estapar.domain.car.exit.CarExitService
+import com.estapar.domain.car.logging.exit.CarExitService
 import com.estapar.domain.car.logging.GarageLoggingService
 import com.estapar.domain.car.park.ParkingService
 import org.springframework.stereotype.Service
